Adding Memcache compat for Finagle & Util upgrade
