+++ storehaus-memcache/src/main/scala/com/twitter/storehaus/memcache/MemcacheStore.scala
@@ -20,14 +20,12 @@ import com.twitter.algebird.Semigroup
 import com.twitter.bijection.{ Codec, Injection }
 import com.twitter.bijection.netty.Implicits._
 import com.twitter.conversions.time._
-import com.twitter.finagle.builder.ClientBuilder
-import com.twitter.finagle.memcached.KetamaClientBuilder
-import com.twitter.finagle.memcached.protocol.text.Memcached
 import com.twitter.finagle.netty3.{BufChannelBuffer, ChannelBufferBuf}
 import com.twitter.util.{ Duration, Future, Time }
 import com.twitter.finagle.memcached.Client
 import com.twitter.storehaus.{ FutureOps, Store, WithPutTtl }
 import com.twitter.storehaus.algebra.MergeableStore
+import com.twitter.storehaus.memcache.compat.MemcacheCompatClient
 import org.jboss.netty.buffer.ChannelBuffer
 
 import Store.enrich
@@ -64,20 +62,8 @@ object MemcacheStore {
     retries: Int = DEFAULT_RETRIES,
     timeout: Duration = DEFAULT_TIMEOUT,
     hostConnectionLimit: Int = DEFAULT_CONNECTION_LIMIT): Client = {
-    val builder = ClientBuilder()
-      .name(name)
-      .retries(retries)
-      .tcpConnectTimeout(timeout)
-      .requestTimeout(timeout)
-      .connectTimeout(timeout)
-      .readerIdleTimeout(timeout)
-      .hostConnectionLimit(hostConnectionLimit)
-      .codec(Memcached())
-
-    KetamaClientBuilder()
-      .clientBuilder(builder)
-      .nodes(nodeString)
-      .build()
+    MemcacheCompatClient.defaultClient(name, nodeString, retries,
+      timeout, hostConnectionLimit)
   }
 
   /**
